chore: large clean up pass
